@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.20.0)
 project(HomeMonitor VERSION 0.1.0 LANGUAGES C CXX)
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 
 add_executable(HomeMonitor main.cpp)
 
